python/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
lightgbm
joblib
pymysql
python-dotenv
